+++ account-keeper-impl/src/main/java/com/dwarfeng/acckeeper/impl/dao/LoginAccountInfoDaoImpl.java
@@ -1,0 +1,128 @@
+package com.dwarfeng.acckeeper.impl.dao;
+
+import com.dwarfeng.acckeeper.sdk.bean.entity.FastJsonLoginAccountInfo;
+import com.dwarfeng.acckeeper.stack.bean.entity.LoginAccountInfo;
+import com.dwarfeng.acckeeper.stack.dao.LoginAccountInfoDao;
+import com.dwarfeng.subgrade.impl.dao.RedisBatchBaseDao;
+import com.dwarfeng.subgrade.impl.dao.RedisEntireLookupDao;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.DaoException;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Repository
+public class LoginAccountInfoDaoImpl implements LoginAccountInfoDao {
+
+    private final RedisBatchBaseDao<StringIdKey, LoginAccountInfo, FastJsonLoginAccountInfo> batchDelegate;
+    private final RedisEntireLookupDao<StringIdKey, LoginAccountInfo, FastJsonLoginAccountInfo> entireLookupDelegate;
+
+    public LoginAccountInfoDaoImpl(
+            RedisBatchBaseDao<StringIdKey, LoginAccountInfo, FastJsonLoginAccountInfo> batchDelegate,
+            RedisEntireLookupDao<StringIdKey, LoginAccountInfo, FastJsonLoginAccountInfo> entireLookupDelegate
+    ) {
+        this.batchDelegate = batchDelegate;
+        this.entireLookupDelegate = entireLookupDelegate;
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public StringIdKey insert(LoginAccountInfo element) throws DaoException {
+        return batchDelegate.insert(element);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void update(LoginAccountInfo element) throws DaoException {
+        batchDelegate.update(element);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void delete(StringIdKey key) throws DaoException {
+        batchDelegate.delete(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean exists(StringIdKey key) throws DaoException {
+        return batchDelegate.exists(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public LoginAccountInfo get(StringIdKey key) throws DaoException {
+        return batchDelegate.get(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public List<StringIdKey> batchInsert(List<LoginAccountInfo> elements) throws DaoException {
+        return batchDelegate.batchInsert(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchUpdate(List<LoginAccountInfo> elements) throws DaoException {
+        batchDelegate.batchUpdate(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDelete(List<StringIdKey> keys) throws DaoException {
+        batchDelegate.batchDelete(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean allExists(List<StringIdKey> keys) throws DaoException {
+        return batchDelegate.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean nonExists(List<StringIdKey> keys) throws DaoException {
+        return batchDelegate.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<LoginAccountInfo> batchGet(List<StringIdKey> keys) throws DaoException {
+        return batchDelegate.batchGet(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<LoginAccountInfo> lookup() throws DaoException {
+        return entireLookupDelegate.lookup();
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<LoginAccountInfo> lookup(PagingInfo pagingInfo) throws DaoException {
+        return entireLookupDelegate.lookup(pagingInfo);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public int lookupCount() throws DaoException {
+        return entireLookupDelegate.lookupCount();
+    }
+}