+++ account-keeper-stack/src/main/java/com/dwarfeng/acckeeper/stack/bean/entity/LoginAccountInfo.java
@@ -1,0 +1,66 @@
+package com.dwarfeng.acckeeper.stack.bean.entity;
+
+import com.dwarfeng.subgrade.stack.bean.entity.Entity;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+
+import java.util.List;
+
+/**
+ * 登陆账户信息。
+ *
+ * @author DwArFeng
+ * @since 1.6.0
+ */
+public class LoginAccountInfo implements Entity<StringIdKey> {
+
+    private static final long serialVersionUID = -5035809484080429449L;
+
+    private StringIdKey key;
+    private List<LongIdKey> stateKeys;
+    private int stateCount;
+
+    public LoginAccountInfo() {
+    }
+
+    public LoginAccountInfo(StringIdKey key, List<LongIdKey> stateKeys, int stateCount) {
+        this.key = key;
+        this.stateKeys = stateKeys;
+        this.stateCount = stateCount;
+    }
+
+    @Override
+    public StringIdKey getKey() {
+        return key;
+    }
+
+    @Override
+    public void setKey(StringIdKey key) {
+        this.key = key;
+    }
+
+    public List<LongIdKey> getStateKeys() {
+        return stateKeys;
+    }
+
+    public void setStateKeys(List<LongIdKey> stateKeys) {
+        this.stateKeys = stateKeys;
+    }
+
+    public int getStateCount() {
+        return stateCount;
+    }
+
+    public void setStateCount(int stateCount) {
+        this.stateCount = stateCount;
+    }
+
+    @Override
+    public String toString() {
+        return "LoginAccountInfo{" +
+                "key=" + key +
+                ", stateKeys=" + stateKeys +
+                ", stateCount=" + stateCount +
+                '}';
+    }
+}